feat(auth): add NextAuth (Google) + Prisma SQLite, marketing page at /, move app to /app, gated middleware (flagged), API auth helper, and user-scoped uploads; add Sign out button in app header
- NextAuth (JWT strategy) with Prisma adapter
- Prisma schema for User/Account/Session/VerificationToken + Project (JSON)
- SessionProvider in _app
- Marketing page at / with Google sign-in CTA
- Original app moved to /app
- middleware.ts with NEXT_PUBLIC_REQUIRE_AUTH flag for gating
- /api/upload now requires auth and writes under temp_images/u_<userId>
- Auth helper requireAuth for APIs
- Added Sign out button in /app header
- .env.example updated with auth + DB settings; DB scripts added
- Ignore prisma/*.db files
