+++ .env.example
@@ -23,6 +23,12 @@ NODE_ENV=development
 # Disable Next.js telemetry
 NEXT_TELEMETRY_DISABLED=1
 
+# Require auth gating (set true in staging/prod when ready)
+NEXT_PUBLIC_REQUIRE_AUTH=false
+
+# Enable server persistence for projects (planned)
+PROJECT_SERVER_PERSISTENCE=false
+
 # ========================================
 # STORAGE CONFIGURATION
 # ========================================
@@ -123,3 +129,15 @@ STORAGE_PROVIDER=local
 # AWS_SES_REGION=us-east-1
 # EMAIL_FROM=certificates@yourdomain.com
 # CLEANUP_SECRET_KEY=your-secret-key
+
+# ========================================
+# AUTH + DATABASE
+# ========================================
+# NextAuth (Auth.js) Google OAuth
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+NEXTAUTH_SECRET= # generate: `openssl rand -base64 32`
+NEXTAUTH_URL=http://localhost:3000
+
+# Prisma / SQLite (local file DB)
+DATABASE_URL="file:./prisma/database.db"
