+++ lib/auth/requireAuth.ts
@@ -1,0 +1,13 @@
+import type { NextApiRequest, NextApiResponse } from 'next';
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from '@/pages/api/auth/[...nextauth]';
+
+export async function requireAuth(req: NextApiRequest, res: NextApiResponse) {
+  const session = await getServerSession(req, res, authOptions as any);
+  if (!session || !session.user) {
+    res.status(401).json({ message: 'Unauthorized' });
+    return null;
+  }
+  return session;
+}
+
