+++ package.json
@@ -25,13 +25,19 @@
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:debug": "playwright test --debug",
     "test:e2e:report": "playwright show-report",
-    "test:e2e:install": "playwright install"
+    "test:e2e:install": "playwright install",
+    "prisma": "prisma",
+    "db:push": "prisma db push",
+    "db:migrate": "prisma migrate dev",
+    "db:studio": "prisma studio"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.842.0",
     "@aws-sdk/client-ses": "^3.840.0",
     "@aws-sdk/s3-request-presigner": "^3.842.0",
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@pdf-lib/fontkit": "^1.1.1",
+    "@prisma/client": "^6.15.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@types/archiver": "^6.0.3",
     "@types/jszip": "^3.4.0",
@@ -49,6 +55,7 @@
     "lucide-react": "^0.441.0",
     "mime-types": "^3.0.1",
     "next": "^15.0.3",
+    "next-auth": "^4.24.11",
     "pdf-lib": "^1.17.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -57,7 +64,8 @@
     "resend": "^4.6.0",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
-    "uuid": "^11.1.0"
+    "uuid": "^11.1.0",
+    "zod": "^4.1.5"
   },
   "devDependencies": {
     "@playwright/test": "^1.54.1",
@@ -77,6 +85,7 @@
     "jest-environment-jsdom": "^30.0.0",
     "node-mocks-http": "^1.17.2",
     "postcss": "^8",
+    "prisma": "^6.15.0",
     "tailwindcss": "^3.4.1",
     "ts-jest": "^29.4.0",
     "ts-loader": "^9.5.2",
