+++ pages/app.tsx
@@ -1,0 +1,1126 @@
+"use client";
+
+import { useState, useMemo, useCallback, useEffect } from "react";
+import Image from "next/image";
+import { Button } from "@/components/ui/button";
+import { useTable, Column } from "react-table";
+import { useTableData } from "@/hooks/useTableData";
+import type { TableData } from "@/types/certificate";
+import { useKeyboardShortcuts } from "@/hooks/useKeyboardShortcuts";
+import { usePreview } from "@/hooks/usePreview";
+import { useFileUpload } from "@/hooks/useFileUpload";
+import { usePositioning } from "@/hooks/usePositioning";
+import { useDragAndDrop } from "@/hooks/useDragAndDrop";
+import { useEmailConfig } from "@/hooks/useEmailConfig";
+import { usePdfGeneration } from "@/hooks/usePdfGeneration";
+import { useProgressivePdfGeneration } from "@/hooks/useProgressivePdfGeneration";
+import { useClientPdfGeneration } from "@/hooks/useClientPdfGeneration";
+import { SPLIT_BUTTON_THEME } from "@/utils/constants";
+import {
+  SkipBack,
+  ChevronLeft,
+  ChevronRight,
+  SkipForward,
+  Save,
+  FileUp,
+  FolderOpen,
+  RefreshCw
+} from "lucide-react";
+import { CertificatePreview } from "@/components/CertificatePreview";
+import { DataPanelWithSearch } from "@/components/panels/DataPanelWithSearch";
+import { FormattingPanel } from "@/components/panels/FormattingPanel";
+import { EmailConfigPanel } from "@/components/panels/EmailConfigPanel";
+import { PdfGenerationModal } from "@/components/modals/PdfGenerationModal";
+import { IndividualPdfsModal } from "@/components/modals/IndividualPdfsModal";
+import { ConfirmationModals } from "@/components/modals/ConfirmationModals";
+import { SaveProjectModal } from "@/components/modals/SaveProjectModal";
+import { LoadProjectModal } from "@/components/modals/LoadProjectModal";
+import { NewProjectModal } from "@/components/modals/NewProjectModal";
+import { ErrorModal } from "@/components/ui/error-alert";
+import { MobileWarningScreen } from "@/components/MobileWarningScreen";
+import { useMobileDetection } from "@/hooks/useMobileDetection";
+import { COLORS, GRADIENTS } from "@/utils/styles";
+import { getImageAverageLuminance, getReadableTextColorForLuminance } from "@/utils/imageAnalysis";
+import { SplitButton } from "@/components/ui/split-button";
+import { useToast, ToastContainer } from "@/components/ui/toast";
+import { useProjectAutosave } from "@/hooks/useProjectAutosave";
+import { useSessionAutosave } from "@/hooks/useSessionAutosave";
+import { useDevMode } from "@/hooks/useDevMode";
+import { DevModeControls } from "@/components/DevModeControls";
+import { usePdfGenerationMethods } from "@/hooks/usePdfGenerationMethods";
+import { useProjectManagement } from "@/hooks/useProjectManagement";
+import { useOnboarding } from "@/hooks/useOnboarding";
+import { OnboardingModal } from "@/components/modals/OnboardingModal";
+import { HelpCircle } from "lucide-react";
+import { useSession, signOut } from 'next-auth/react';
+
+export default function HomePage() {
+  // ============================================================================
+  // MOBILE DETECTION
+  // ============================================================================
+
+  const { isMobile, isLoading: isMobileLoading } = useMobileDetection();
+  const [forceMobileAccess, setForceMobileAccess] = useState(false);
+  const { data: session } = useSession();
+
+  // ============================================================================
+  // ONBOARDING & TUTORIAL
+  // ============================================================================
+  
+  const {
+    showOnboarding,
+    hasSeenOnboarding,
+    startTour,
+    skipOnboarding,
+    setShowOnboarding
+  } = useOnboarding();
+
+  // Check if user is completely new (show onboarding modal)
+  useEffect(() => {
+    if (!hasSeenOnboarding && !isMobileLoading && !isMobile) {
+      setShowOnboarding(true);
+    }
+  }, [hasSeenOnboarding, isMobileLoading, isMobile, setShowOnboarding]);
+
+
+  // ============================================================================
+  // STATE & DATA MANAGEMENT
+  // ============================================================================
+
+  const isDevelopment = process.env.NODE_ENV === "development";
+
+  // Dev mode state (defined early to avoid dependency issues)
+  const [devMode, setDevMode] = useState<boolean>(false);
+  const [emailTemplate, setEmailTemplate] = useState<string>("");
+  const [numTestEmails, setNumTestEmails] = useState<number>(10);
+
+  // Table data management via custom hook
+  const {
+    tableData,
+    tableInput,
+    isFirstRowHeader,
+    useCSVMode,
+    detectedEmailColumn,
+    isProcessingData,
+    handleTableDataChange,
+    handleHeaderToggle,
+    handleCSVModeToggle,
+    loadPresetData,
+    clearData,
+    loadSessionData
+  } = useTableData();
+
+  // ============================================================================
+  // CUSTOM HOOKS FOR FEATURE MANAGEMENT
+  // ============================================================================
+
+  // Selected field state (needs to be before usePositioning)
+  const [selectedField, setSelectedField] = useState<string | null>(null);
+  
+  // Positioning hook
+  const { positions, setPositions, changeAlignment, clearPositions } =
+    usePositioning({ tableData, setSelectedField });
+  const [activeTab, setActiveTab] = useState<"data" | "formatting" | "email">(
+    "data"
+  );
+  const [showAppliedMessage, setShowAppliedMessage] = useState<boolean>(false);
+  const [selectedNamingColumn, setSelectedNamingColumn] = useState<string>("");
+  const [showResetFieldModal, setShowResetFieldModal] =
+    useState<boolean>(false);
+  const [showClearAllModal, setShowClearAllModal] = useState<boolean>(false);
+  const [hasInputFocus, setHasInputFocus] = useState<boolean>(false);
+
+  // Drag and drop hook
+  const {
+    isDragging,
+    dragInfo,
+    showCenterGuide,
+    setShowCenterGuide,
+    handlePointerDown,
+    handlePointerUp,
+    clearDragState
+  } = useDragAndDrop({
+    positions,
+    setPositions,
+    setSelectedField,
+    setActiveTab
+  });
+
+  // Switch away from email tab if no email column detected
+  useEffect(() => {
+    console.log(
+      "👀 Email tab visibility check - detectedEmailColumn:",
+      detectedEmailColumn,
+      "activeTab:",
+      activeTab
+    );
+    if (!detectedEmailColumn && activeTab === "email") {
+      setActiveTab("data");
+    }
+  }, [detectedEmailColumn, activeTab]);
+
+  // Switch away from formatting tab if no table data
+  useEffect(() => {
+    if (tableData.length === 0 && activeTab === "formatting") {
+      setActiveTab("data");
+    }
+  }, [tableData.length, activeTab]);
+
+  // Preview navigation hook
+  const {
+    currentPreviewIndex,
+    goToFirst,
+    goToPrevious,
+    goToNext,
+    goToLast,
+    setCurrentPreviewIndex
+  } = usePreview(tableData.length, Object.keys(positions).length);
+
+  // File upload hook
+  const {
+    uploadedFile,
+    uploadedFileUrl,
+    isLoading,
+    isDraggingFile,
+    setUploadedFile,
+    setUploadedFileUrl,
+    handleFileUpload,
+    handleDragOver,
+    handleDragLeave,
+    handleFileDrop,
+    clearFile,
+    uploadError,
+    clearError,
+    localBlobUrl,
+    uploadToServer
+  } = useFileUpload();
+
+  // Adapt default text color to background tone when a new image is loaded
+  useEffect(() => {
+    (async () => {
+      if (!uploadedFileUrl || tableData.length === 0) return;
+      try {
+        const lum = await getImageAverageLuminance(uploadedFileUrl);
+        const autoColor = getReadableTextColorForLuminance(lum); // '#000000' or '#ffffff'
+
+        setPositions((prev) => {
+          const next = { ...prev };
+          let changed = false;
+          const defaultBlack = '#000000';
+          const defaultWhite = '#ffffff';
+
+          // Only set auto color for fields that are at default color
+          Object.keys(tableData[0] || {}).forEach((key) => {
+            const current = next[key];
+            if (!current) return;
+            const currentColor = (current.color || '').toLowerCase();
+            if (current.color === undefined || currentColor === defaultBlack || currentColor === '') {
+              // untouched or default black -> apply auto color
+              next[key] = { ...current, color: autoColor };
+              changed = true;
+            } else if ((currentColor === defaultWhite || currentColor === defaultBlack) && currentColor !== autoColor) {
+              // If a prior auto applied simple black/white exists, keep adapting
+              next[key] = { ...current, color: autoColor };
+              changed = true;
+            }
+          });
+          return changed ? next : prev;
+        });
+      } catch {
+        // Ignore analysis errors silently
+      }
+    })();
+  }, [uploadedFileUrl, tableData, setPositions]);
+
+
+  // PDF generation hook (must come after file upload hook)
+  const {
+    isGenerating,
+    isGeneratingIndividual,
+    generatedPdfUrl,
+    individualPdfsData,
+    generatePdf,
+    generateIndividualPdfs,
+    handleDownloadPdf,
+    setGeneratedPdfUrl,
+    setIndividualPdfsData,
+    clearPdfData
+  } = usePdfGeneration({
+    tableData,
+    positions,
+    uploadedFile,
+    selectedNamingColumn,
+    setSelectedNamingColumn
+  });
+
+  // Progressive PDF generation hook
+  const {
+    isGenerating: isProgressiveGenerating,
+    progress: progressivePdfProgress,
+    results: progressivePdfResults,
+    error: progressivePdfError,
+    startProgressiveGeneration,
+    pauseGeneration,
+    resumeGeneration,
+    cancelGeneration,
+    clearResults
+  } = useProgressivePdfGeneration({
+    tableData,
+    positions,
+    uploadedFile,
+    selectedNamingColumn,
+    setSelectedNamingColumn
+  });
+
+  // Client-side PDF generation hook (always enabled)
+  const {
+    isClientSupported,
+    isGenerating: isClientGenerating,
+    isGeneratingIndividual: isClientGeneratingIndividual,
+    generatedPdfUrl: clientGeneratedPdfUrl,
+    individualPdfsData: clientIndividualPdfsData,
+    generatePdf: generateClientPdf,
+    generateIndividualPdfs: generateClientIndividualPdfs,
+    handleDownloadPdf: handleClientDownloadPdf,
+    clearPdfData: clearClientPdfData
+  } = useClientPdfGeneration({
+    tableData,
+    positions,
+    uploadedFile,
+    uploadedFileUrl,
+    localBlobUrl,
+    selectedNamingColumn,
+    setSelectedNamingColumn,
+    enabled: true // Always enabled, client-side is default
+  });
+
+  // Mark variables as intentionally unused (kept for future features)
+  void isClientGenerating;
+  void handleClientDownloadPdf;
+
+  // PDF generation methods hook
+  const { handleGeneratePdf, handleGenerateIndividualPdfs } = usePdfGenerationMethods({
+    isDevelopment,
+    devMode,
+    isClientSupported,
+    tableData,
+    localBlobUrl,
+    uploadedFileUrl,
+    uploadedFile,
+    generatePdf,
+    generateIndividualPdfs,
+    startProgressiveGeneration,
+    setGeneratedPdfUrl,
+    setIndividualPdfsData,
+    generateClientPdf,
+    generateClientIndividualPdfs,
+    clientGeneratedPdfUrl,
+    clientIndividualPdfsData,
+    uploadToServer
+  });
+
+  // Email configuration hook (must come after PDF generation hook)
+  const {
+    emailConfig,
+    setEmailConfig,
+    emailSendingStatus,
+    sendCertificateEmail,
+    hasEmailColumn
+  } = useEmailConfig({
+    detectedEmailColumn,
+    tableData,
+    individualPdfsData
+  });
+
+  // Dev mode hook (must come after file upload, PDF generation, and email config hooks)
+  const { handleDevModeToggle, handleEmailTemplateUpdate } = useDevMode({
+    isDevelopment,
+    devMode,
+    setDevMode,
+    emailTemplate,
+    setEmailTemplate,
+    numTestEmails,
+    setNumTestEmails,
+    loadPresetData,
+    clearData,
+    clearFile,
+    clearPositions,
+    clearPdfData,
+    setUploadedFileUrl,
+    setUploadedFile,
+    setEmailConfig
+  });
+
+  // Toast notifications
+  const { toasts, showToast, hideToast } = useToast();
+
+  // Early autosave hook for manual save function
+  const { manualSave: baseManualSave } = useProjectAutosave({
+    positions,
+    columns: Object.keys(tableData[0] || {}),
+    emailConfig,
+    certificateImageUrl: uploadedFileUrl,
+    certificateFilename: uploadedFile as string | null,
+    tableData,
+    onAutosave: () => {
+      // Silent autosave - no toast notification
+      console.log("Project autosaved");
+    },
+    enabled: false, // Base hook is disabled, only used for manual save
+    currentProjectId: null,
+    currentProjectName: null
+  });
+
+  // Project management hook
+  const {
+    currentProjectName,
+    currentProjectId,
+    hasManuallySaved,
+    showSaveProjectModal,
+    showLoadProjectModal,
+    showNewProjectModal,
+    setShowSaveProjectModal,
+    setShowLoadProjectModal,
+    setShowNewProjectModal,
+    handleLoadProject,
+    handleSaveProjectSuccess,
+    handleSaveToCurrentProject,
+    handleNewProject,
+    confirmNewProject
+  } = useProjectManagement({
+    setPositions,
+    setEmailConfig,
+    setUploadedFileUrl,
+    setUploadedFile,
+    loadSessionData,
+    uploadToServer,
+    showToast,
+    manualSave: baseManualSave,
+    uploadedFileUrl,
+    uploadedFile,
+    positions,
+    emailConfig,
+    tableData,
+    clearFile,
+    clearPositions,
+    clearDragState,
+    clearData
+  });
+
+  // Session data autosave hook (only enabled after manual save)
+  useSessionAutosave({
+    tableData,
+    tableInput,
+    isFirstRowHeader,
+    useCSVMode,
+    enabled: hasManuallySaved && tableData.length > 0
+  });
+
+  // Project autosave hook - actually enables autosave after manual save
+  useProjectAutosave({
+    positions,
+    columns: Object.keys(tableData[0] || {}),
+    emailConfig,
+    certificateImageUrl: uploadedFileUrl,
+    certificateFilename: uploadedFile as string | null,
+    tableData,
+    onAutosave: () => {
+      // Silent autosave - no toast notification
+      console.log("Project autosaved");
+    },
+    enabled: hasManuallySaved,
+    currentProjectId,
+    currentProjectName
+  });
+
+  // ============================================================================
+  // EVENT HANDLERS & BUSINESS LOGIC
+  // ============================================================================
+
+  // Handle escape key press to close all modals
+  const handleEscapePressed = useCallback(() => {
+    setGeneratedPdfUrl(null);
+    setIndividualPdfsData(null);
+    setShowResetFieldModal(false);
+    setShowClearAllModal(false);
+  }, [setGeneratedPdfUrl, setIndividualPdfsData]); // Include setters for completeness
+
+  // Track input focus state globally
+  useEffect(() => {
+    const handleFocusIn = (event: FocusEvent) => {
+      const target = event.target as HTMLElement;
+      if (
+        target?.tagName === 'INPUT' ||
+        target?.tagName === 'TEXTAREA' ||
+        target?.getAttribute('contenteditable') === 'true'
+      ) {
+        setHasInputFocus(true);
+      }
+    };
+
+    const handleFocusOut = (event: FocusEvent) => {
+      const target = event.target as HTMLElement;
+      if (
+        target?.tagName === 'INPUT' ||
+        target?.tagName === 'TEXTAREA' ||
+        target?.getAttribute('contenteditable') === 'true'
+      ) {
+        setHasInputFocus(false);
+      }
+    };
+
+    document.addEventListener('focusin', handleFocusIn);
+    document.addEventListener('focusout', handleFocusOut);
+
+    return () => {
+      document.removeEventListener('focusin', handleFocusIn);
+      document.removeEventListener('focusout', handleFocusOut);
+    };
+  }, []);
+
+  // Keyboard shortcuts hook
+  useKeyboardShortcuts({
+    selectedField,
+    isDragging,
+    positions,
+    setPositions,
+    onEscapePressed: handleEscapePressed
+  });
+
+
+  // ============================================================================
+  // TABLE CONFIGURATION
+  // ============================================================================
+
+  const columns: Column<TableData>[] = useMemo(
+    () =>
+      tableData.length > 0
+        ? Object.keys(tableData[0]).map((key, index) => ({
+            Header: key || `Column ${index + 1}`, // Display name for blank headers
+            accessor: key || `_column_${index}` // Ensure valid accessor for React Table
+          }))
+        : [],
+    [tableData]
+  );
+
+  const { headerGroups, rows, prepareRow } = useTable({
+    columns,
+    data: tableData
+  });
+
+  // ============================================================================
+  // MOBILE DETECTION LOGIC
+  // ============================================================================
+
+  // Listen for force mobile access event
+  useEffect(() => {
+    const handleForceMobileAccess = () => {
+      setForceMobileAccess(true);
+    };
+
+    window.addEventListener("forceMobileAccess", handleForceMobileAccess);
+    return () =>
+      window.removeEventListener("forceMobileAccess", handleForceMobileAccess);
+  }, []);
+
+  // Show mobile warning if on mobile and haven't forced access
+  if (!isMobileLoading && isMobile && !forceMobileAccess) {
+    return <MobileWarningScreen />;
+  }
+
+  // ============================================================================
+  // RENDER
+  // ============================================================================
+
+  return (
+    <div
+      className="flex flex-col h-screen"
+      style={{ backgroundColor: COLORS.background }}>
+      {/* Application Header */}
+      <header
+        className="py-4 px-6"
+        style={{
+          background: GRADIENTS.primary,
+          boxShadow: "0 2px 4px rgba(27, 67, 50, 0.1)"
+        }}>
+        <div className="flex justify-between items-center">
+          <div className="flex items-center gap-4">
+            <div className="flex items-center gap-3">
+              <Image
+                src="/bamboobot-icon.png"
+                alt="Bamboobot"
+                width={32}
+                height={32}
+                className="w-8 h-8"
+              />
+              <h1
+                className="text-2xl font-bold"
+                style={{ color: COLORS.amber }}>
+                Bamboobot
+              </h1>
+            </div>
+            {/* Dev Mode Controls - Only visible in development */}
+            <DevModeControls
+              isDevelopment={isDevelopment}
+              devMode={devMode}
+              handleDevModeToggle={handleDevModeToggle}
+              emailTemplate={emailTemplate}
+              setEmailTemplate={setEmailTemplate}
+              numTestEmails={numTestEmails}
+              setNumTestEmails={setNumTestEmails}
+              handleEmailTemplateUpdate={handleEmailTemplateUpdate}
+            />
+          </div>
+          <div className="flex gap-3">
+            {/* Help/Tutorial Button */}
+            <Button
+              onClick={() => setShowOnboarding(true)}
+              variant="outline"
+              className="gap-2 bg-white/10 border-white/20 text-white hover:bg-white/20"
+            >
+              <HelpCircle className="w-4 h-4" />
+              Tutorial
+            </Button>
+            
+            {/* Projects Split Button */}
+            <SplitButton
+              data-tour="save-project"
+              label={
+                currentProjectName 
+                  ? `Save to "${currentProjectName}"` 
+                  : "Save project"
+              }
+              onClick={
+                hasManuallySaved
+                  ? handleSaveToCurrentProject // Save to current project
+                  : () => setShowSaveProjectModal(true)
+              }
+              disabled={false}
+              variant="primary" // Always use primary variant to maintain consistent styling
+              menuItems={[
+                {
+                  label: "New project",
+                  icon: <FileUp className="h-4 w-4" />,
+                  onClick: handleNewProject,
+                  disabled:
+                    !uploadedFileUrl && Object.keys(positions).length === 0
+                },
+                {
+                  label: "Projects...",
+                  icon: <FolderOpen className="h-4 w-4" />,
+                  onClick: () => setShowLoadProjectModal(true)
+                },
+                {
+                  label: "Save as new project",
+                  icon: <Save className="h-4 w-4" />,
+                  onClick: () => setShowSaveProjectModal(true),
+                  disabled:
+                    !uploadedFileUrl || Object.keys(positions).length === 0
+                }
+              ]}
+              gradientClass={SPLIT_BUTTON_THEME.projects.gradient}
+              dropdownColor={SPLIT_BUTTON_THEME.projects.dropdownColor}
+              dropdownHoverColor={
+                SPLIT_BUTTON_THEME.projects.dropdownHoverColor
+              }
+            />
+
+            {/* Generate Split Button */}
+            <SplitButton
+              data-tour="generate-pdf"
+              label="Generate"
+              onClick={() => {
+                handleGenerateIndividualPdfs();
+              }}
+              menuItems={[
+                {
+                  label: "Individual PDFs",
+                  onClick: () => {
+                    handleGenerateIndividualPdfs();
+                  },
+                  disabled:
+                    !uploadedFile ||
+                    isGenerating ||
+                    isGeneratingIndividual ||
+                    isProgressiveGenerating ||
+                    tableData.length === 0
+                },
+                {
+                  label: "Single PDF",
+                  onClick: () => handleGeneratePdf(),
+                  disabled:
+                    !uploadedFile ||
+                    isGenerating ||
+                    isGeneratingIndividual ||
+                    tableData.length === 0
+                },
+                // Server-side options only in dev mode
+                ...(isDevelopment && devMode ? [
+                  {
+                    label: "──────────",
+                    onClick: () => {},
+                    disabled: true
+                  },
+                  {
+                    label: "Individual PDFs (Server)",
+                    onClick: () => {
+                      handleGenerateIndividualPdfs(true);
+                    },
+                    disabled:
+                      !uploadedFile ||
+                      isGenerating ||
+                      isGeneratingIndividual ||
+                      isProgressiveGenerating ||
+                      tableData.length === 0
+                  },
+                  {
+                    label: "Single PDF (Server)",
+                    onClick: () => handleGeneratePdf(true),
+                    disabled:
+                      !uploadedFile ||
+                      isGenerating ||
+                      isGeneratingIndividual ||
+                      tableData.length === 0
+                  }
+                ] : [])
+              ]}
+              disabled={
+                !uploadedFile ||
+                isGenerating ||
+                isGeneratingIndividual ||
+                isProgressiveGenerating ||
+                tableData.length === 0
+              }
+              gradientClass={SPLIT_BUTTON_THEME.generate.gradient}
+              dropdownColor={SPLIT_BUTTON_THEME.generate.dropdownColor}
+              dropdownHoverColor={
+                SPLIT_BUTTON_THEME.generate.dropdownHoverColor
+              }
+            />
+            {session && (
+              <Button
+                onClick={() => signOut()}
+                variant="outline"
+                className="bg-white/10 border-white/20 text-white hover:bg-white/20"
+                title={session.user?.email || 'Sign out'}
+              >
+                Sign out
+              </Button>
+            )}
+          </div>
+        </div>
+      </header>
+
+      {/* Main Content Area */}
+      <main className="flex-1 grid grid-cols-[60%_40%] gap-6 p-6">
+        {/* Certificate Preview Section */}
+        <div className="bg-card p-4 rounded-lg shadow" data-tour="certificate-preview">
+          <CertificatePreview
+            uploadedFileUrl={uploadedFileUrl}
+            isLoading={isLoading}
+            tableData={tableData}
+            currentPreviewIndex={currentPreviewIndex}
+            positions={positions}
+            selectedField={selectedField}
+            setSelectedField={setSelectedField}
+            isDragging={isDragging}
+            dragInfo={dragInfo}
+            showCenterGuide={showCenterGuide}
+            handlePointerDown={handlePointerDown}
+            handlePointerUp={handlePointerUp}
+            setShowCenterGuide={setShowCenterGuide}
+            isDraggingFile={isDraggingFile}
+            handleDragOver={handleDragOver}
+            handleDragLeave={handleDragLeave}
+            handleFileDrop={handleFileDrop}
+            handleFileUpload={handleFileUpload}
+          />
+
+          {/* Toolbar */}
+          {uploadedFileUrl && (
+            <div className="mt-4 p-3 bg-gray-50 rounded-lg border border-gray-200">
+              <div className="flex justify-between items-center">
+                <div className="flex gap-2">
+                  <SplitButton
+                    label="Remove Background"
+                    onClick={() => {
+                      clearFile();
+                      clearDragState();
+                    }}
+                    menuItems={[
+                      {
+                        label: "Replace Background",
+                        icon: <RefreshCw className="h-4 w-4" />,
+                        onClick: () => {
+                          const fileInput = document.getElementById('replace-file-upload') as HTMLInputElement;
+                          fileInput?.click();
+                        }
+                      }
+                    ]}
+                    variant="secondary"
+                    className="bg-white"
+                    dropUp={true}
+                  />
+                  <input
+                    id="replace-file-upload"
+                    type="file"
+                    onChange={handleFileUpload}
+                    accept="image/jpeg,image/png"
+                    className="sr-only"
+                  />
+                </div>
+
+                {/* Arrow key hint - center aligned */}
+                {selectedField && (
+                  <div className="flex items-center justify-center text-xs text-gray-500">
+                    <span className={hasInputFocus ? "opacity-50" : ""}>
+                      {hasInputFocus 
+                        ? "Arrow key nudging disabled while editing"
+                        : "Use arrow keys to nudge selected text (Shift for larger steps)"}
+                    </span>
+                  </div>
+                )}
+
+                {/* Navigation buttons - right aligned */}
+                {tableData.length > 0 && (
+                  <div className="flex items-center gap-2" data-tour="navigation">
+                    <span className="text-sm text-gray-600 font-medium">
+                      {currentPreviewIndex + 1} of {tableData.length}
+                    </span>
+                    <div className="flex gap-1">
+                      <Button
+                        variant="outline"
+                        size="sm"
+                        disabled={currentPreviewIndex === 0}
+                        className={
+                          currentPreviewIndex === 0
+                            ? "text-gray-400 border-gray-300 px-2"
+                            : "text-gray-700 border-gray-400 px-2 hover:bg-gray-50"
+                        }
+                        title="First entry"
+                        onClick={goToFirst}>
+                        <SkipBack className="h-4 w-4" />
+                      </Button>
+                      <Button
+                        variant="outline"
+                        size="sm"
+                        disabled={currentPreviewIndex === 0}
+                        className={
+                          currentPreviewIndex === 0
+                            ? "text-gray-400 border-gray-300 px-2"
+                            : "text-gray-700 border-gray-400 px-2 hover:bg-gray-50"
+                        }
+                        title="Previous entry"
+                        onClick={goToPrevious}>
+                        <ChevronLeft className="h-4 w-4" />
+                      </Button>
+                      <Button
+                        variant="outline"
+                        size="sm"
+                        disabled={currentPreviewIndex === tableData.length - 1}
+                        className={
+                          currentPreviewIndex === tableData.length - 1
+                            ? "text-gray-400 border-gray-300 px-2"
+                            : "text-gray-700 border-gray-400 px-2 hover:bg-gray-50"
+                        }
+                        title="Next entry"
+                        onClick={goToNext}>
+                        <ChevronRight className="h-4 w-4" />
+                      </Button>
+                      <Button
+                        variant="outline"
+                        size="sm"
+                        disabled={currentPreviewIndex === tableData.length - 1}
+                        className={
+                          currentPreviewIndex === tableData.length - 1
+                            ? "text-gray-400 border-gray-300 px-2"
+                            : "text-gray-700 border-gray-400 px-2 hover:bg-gray-50"
+                        }
+                        title="Last entry"
+                        onClick={goToLast}>
+                        <SkipForward className="h-4 w-4" />
+                      </Button>
+                    </div>
+                  </div>
+                )}
+              </div>
+            </div>
+          )}
+        </div>
+        {/* Control Panel Section */}
+        <div className="bg-card p-4 rounded-lg shadow mr-6">
+          {/* Tab Navigation */}
+          <div className="flex mb-4 bg-gray-100 rounded-lg p-1 gap-1">
+            <button
+              onClick={() => setActiveTab("data")}
+              className={`px-4 py-2 text-sm font-medium rounded-md transition-all flex-1 text-center`}
+              data-tour="data-tab"
+              style={{
+                backgroundColor:
+                  activeTab === "data" ? COLORS.tabActive : COLORS.tabInactive,
+                color:
+                  activeTab === "data" ? COLORS.tabTextActive : COLORS.tabText
+              }}>
+              Data
+            </button>
+            {tableData.length > 0 && (
+              <button
+                onClick={() => setActiveTab("formatting")}
+                className={`px-4 py-2 text-sm font-medium rounded-md transition-all flex-1 text-center`}
+                data-tour="formatting-tab"
+                style={{
+                  backgroundColor:
+                    activeTab === "formatting"
+                      ? COLORS.tabActive
+                      : COLORS.tabInactive,
+                  color:
+                    activeTab === "formatting"
+                      ? COLORS.tabTextActive
+                      : COLORS.tabText
+                }}>
+                Formatting
+              </button>
+            )}
+            {detectedEmailColumn && (
+              <button
+                onClick={() => {
+                  console.log("📧 Email tab clicked");
+                  setActiveTab("email");
+                }}
+                className="px-4 py-2 text-sm font-medium rounded-md transition-all flex-1 text-center"
+                data-tour="email-tab"
+                style={{
+                  backgroundColor:
+                    activeTab === "email"
+                      ? COLORS.tabActive
+                      : COLORS.tabInactive,
+                  color:
+                    activeTab === "email"
+                      ? COLORS.tabTextActive
+                      : COLORS.tabText
+                }}>
+                Email
+                {detectedEmailColumn &&
+                  !["email", "e-mail", "mail"].includes(
+                    detectedEmailColumn.toLowerCase()
+                  ) && (
+                    <span
+                      className={`ml-1 px-1.5 py-0.5 text-xs ${
+                        activeTab === "email"
+                          ? "text-amber-600"
+                          : "text-green-800"
+                      }`}
+                      style={{
+                        backgroundColor:
+                          activeTab === "email"
+                            ? COLORS.amber
+                            : COLORS.successLight,
+                        borderRadius: "4px"
+                      }}>
+                      {detectedEmailColumn}
+                    </span>
+                  )}
+              </button>
+            )}
+          </div>
+
+          {/* Tab Content */}
+          {activeTab === "data" && (
+            <DataPanelWithSearch
+              tableInput={tableInput}
+              handleTableDataChange={handleTableDataChange}
+              isFirstRowHeader={isFirstRowHeader}
+              handleHeaderToggle={handleHeaderToggle}
+              useCSVMode={useCSVMode}
+              handleCSVModeToggle={handleCSVModeToggle}
+              tableData={tableData}
+              headerGroups={headerGroups}
+              rows={rows}
+              prepareRow={prepareRow}
+              detectedEmailColumn={detectedEmailColumn}
+              currentPreviewIndex={currentPreviewIndex}
+              setCurrentPreviewIndex={setCurrentPreviewIndex}
+              isProcessing={isProcessingData}
+            />
+          )}
+
+          {activeTab === "formatting" && (
+            <FormattingPanel
+              selectedField={selectedField}
+              setSelectedField={setSelectedField}
+              positions={positions}
+              setPositions={setPositions}
+              tableData={tableData}
+              changeAlignment={changeAlignment}
+              showAppliedMessage={showAppliedMessage}
+              setShowAppliedMessage={setShowAppliedMessage}
+              setShowResetFieldModal={setShowResetFieldModal}
+              setShowClearAllModal={setShowClearAllModal}
+              setShowCenterGuide={setShowCenterGuide}
+            />
+          )}
+
+          {activeTab === "email" && (
+            <EmailConfigPanel
+              detectedEmailColumn={detectedEmailColumn}
+              emailConfig={emailConfig}
+              setEmailConfig={setEmailConfig}
+              selectedField={selectedField}
+              setSelectedField={setSelectedField}
+              setShowCenterGuide={setShowCenterGuide}
+            />
+          )}
+        </div>
+      </main>
+
+      {/* Modal Components */}
+      <PdfGenerationModal
+        isGenerating={isGenerating}
+        generatedPdfUrl={generatedPdfUrl}
+        handleDownloadPdf={handleDownloadPdf}
+        setGeneratedPdfUrl={setGeneratedPdfUrl}
+        onClose={() => setGeneratedPdfUrl(null)}
+      />
+
+      <IndividualPdfsModal
+        isGeneratingIndividual={
+          isGeneratingIndividual || isProgressiveGenerating || isClientGeneratingIndividual
+        }
+        individualPdfsData={
+          individualPdfsData || clientIndividualPdfsData ||
+          (progressivePdfResults && progressivePdfResults.files.length > 0
+            ? progressivePdfResults.files.map((file) => ({
+                filename: file.filename,
+                url: file.path,
+                originalIndex: file.index
+              }))
+            : null)
+        }
+        tableData={tableData}
+        selectedNamingColumn={selectedNamingColumn}
+        setSelectedNamingColumn={setSelectedNamingColumn}
+        emailSendingStatus={emailSendingStatus}
+        hasEmailColumn={hasEmailColumn}
+        emailConfig={emailConfig}
+        sendCertificateEmail={sendCertificateEmail}
+        setIndividualPdfsData={setIndividualPdfsData}
+        detectedEmailColumn={detectedEmailColumn}
+        // Progressive generation props
+        isProgressiveMode={isProgressiveGenerating || !!progressivePdfProgress}
+        progressiveProgress={progressivePdfProgress}
+        progressiveError={progressivePdfError}
+        onProgressivePause={pauseGeneration}
+        onProgressiveResume={resumeGeneration}
+        onProgressiveCancel={cancelGeneration}
+        onClose={() => {
+          setIndividualPdfsData(null);
+          setSelectedNamingColumn("");
+          if (isProgressiveGenerating || progressivePdfProgress) {
+            clearResults();
+          }
+          if (clientIndividualPdfsData) {
+            clearClientPdfData();
+          }
+        }}
+      />
+
+      <ConfirmationModals
+        showResetFieldModal={showResetFieldModal}
+        setShowResetFieldModal={setShowResetFieldModal}
+        showClearAllModal={showClearAllModal}
+        setShowClearAllModal={setShowClearAllModal}
+        selectedField={selectedField}
+        positions={positions}
+        setPositions={setPositions}
+        tableData={tableData}
+      />
+
+      {/* Project Modals */}
+      <SaveProjectModal
+        isOpen={showSaveProjectModal}
+        onClose={() => setShowSaveProjectModal(false)}
+        positions={positions}
+        columns={Object.keys(tableData[0] || {})}
+        tableData={tableData}
+        emailConfig={emailConfig}
+        certificateImageUrl={uploadedFileUrl || undefined}
+        certificateFilename={(uploadedFile as string) || undefined}
+        onSaveSuccess={handleSaveProjectSuccess}
+        onManualSave={async (projectName) => {
+          let finalUrl = uploadedFileUrl;
+          let finalFilename = uploadedFile as string;
+          
+          // Ensure file is uploaded to server before saving project
+          // Check if we're using a blob URL (local file not yet uploaded)
+          if (uploadedFileUrl?.startsWith('blob:')) {
+            console.log('Uploading file to server before saving project...');
+            try {
+              const uploadResult = await uploadToServer();
+              if (uploadResult) {
+                finalUrl = uploadResult.image;
+                finalFilename = uploadResult.filename;
+                console.log("File uploaded, using server URL:", finalUrl);
+              }
+            } catch (error) {
+              console.error('Failed to upload file before saving:', error);
+              return { success: false, error: 'Failed to upload file to server' };
+            }
+          }
+          return baseManualSave(projectName, finalUrl ?? undefined, finalFilename ?? undefined);
+        }}
+      />
+
+      <LoadProjectModal
+        isOpen={showLoadProjectModal}
+        onClose={() => setShowLoadProjectModal(false)}
+        onLoadProject={handleLoadProject}
+      />
+
+      <NewProjectModal
+        isOpen={showNewProjectModal}
+        onClose={() => setShowNewProjectModal(false)}
+        onConfirm={confirmNewProject}
+        hasUnsavedWork={
+          uploadedFileUrl !== null && Object.keys(positions).length > 0
+        }
+      />
+
+      {/* Toast Container */}
+      <ToastContainer toasts={toasts} onClose={hideToast} />
+
+  
+      {/* Error Modal */}
+      {uploadError && (
+        <ErrorModal
+          open={true}
+          title={uploadError.title}
+          message={uploadError.message}
+          action={uploadError.action}
+          onClose={clearError}
+          onRetry={() => {
+            clearError();
+            document.getElementById("file-upload")?.click();
+          }}
+        />
+      )}
+
+
+      {/* Onboarding Modal */}
+      <OnboardingModal
+        isOpen={showOnboarding}
+        onClose={() => setShowOnboarding(false)}
+        onStartTour={() => {
+          setShowOnboarding(false);
+          // Small delay to ensure modal is closed before starting tour
+          setTimeout(() => {
+            startTour();
+          }, 300);
+        }}
+        onSkip={() => {
+          setShowOnboarding(false);
+          skipOnboarding();
+        }}
+      />
+    </div>
+  );
+}
